Guard clause implementation
Improve readability and overall scalability by implementing simple guard clauses that have been thoroughly tested. More information on guard clauses can be found here: https://www.youtube.com/watch?v=EumXak7TyQ0&t=385s&ab_channel=WebDevSimplified. 

Also, I noticed that in the last event you check if the interaction has autocomplete but then also check if it's one of the 4 commands that use autocomplete. However, would this be necessary as the only autocomplete commands are those 4 so checking if the command is in that list is not needed 🤔? I will leave it to your discretion 😄.